refactor(react): move file-type-utils to react
Relocates isPreviewableImage and isPreviewableSvg to
@self-review/react since they are only consumed by FileSection.tsx.
This eliminates the need for a browser shim when consuming
@self-review/react in Vite-based apps, as @self-review/core no
longer needs to be aliased to avoid its Node.js-only imports.
