+++ packages/core/src/browser.ts
@@ -32,8 +32,5 @@ export type {
 // Diff parsing (pure JS, no Node.js deps)
 export { parseDiff } from './diff-parser';
 
-// File type detection utilities (pure JS, no Node.js deps)
-export { isPreviewableImage, isPreviewableSvg } from './file-type-utils';
-
 // Ignore filter (uses `ignore` package, browser-safe)
 export { createIgnoreFilter } from './ignore-filter';
