+++ packages/core/src/index.ts
@@ -64,5 +64,3 @@ export { createIgnoreFilter } from './ignore-filter';
 // File system utilities
 export { checkWritability } from './fs-utils';
 
-// File type detection utilities
-export { isPreviewableImage, isPreviewableSvg } from './file-type-utils';
